+++ app/assets/javascripts/application.js
@@ -15,6 +15,7 @@
 //= require jquery.autosize
 //= require sjcl
 //= require zeroclipboard
+//= require markdown
 //= require_tree .
 
 function randomString(length) {
@@ -98,9 +99,11 @@ $(document).ready(function(){
   // Show decrypted message
   var decryptedMessage = sjcl.decrypt(password, encryptedRebuilt);
   decryptedMessage = escapeHtml(decryptedMessage);
-  decryptedMessage = decryptedMessage.replace(/\n/g, '<br/>');
-  decryptedMessage = decryptedMessage.replace(/\s/g, '&nbsp;');
+  decryptedMessage = markdown.toHTML(decryptedMessage);
+  // decryptedMessage = decryptedMessage.replace(/\n/g, '<br/>');
+  // decryptedMessage = decryptedMessage.replace(/\s/g, '&nbsp;');
   encryptedTextArea.html(decryptedMessage);
+
 });
 
 function escapeHtml(str) {